Updated Dao structure
