+++ src/main/java/com/globallogic/amcr/persistence/dao/Dao.java
@@ -4,32 +4,32 @@ import java.util.List;
 import java.util.UUID;
 
 /**
- * @param <T> the type of object to be saved
- * @param <R> the type of object to be returned
+ * @param <T> The type of object to be saved
+ * @param <R> The type of object to be returned
  */
 public interface Dao<T, R> {
 
     /**
-     * prepares saves an object to a table
+     * Prepares and saves an object to a table
      *
-     * @param t    the object to be saved
-     * @param uuid the id of the object to be saved
-     * @return returns the saved object
+     * @param t    The object to be saved
+     * @param uuid The id of the object to be saved
+     * @return Returns the saved object
      */
     T save(T t, UUID uuid);
 
     /**
-     * requests a single item from the table with specified id
+     * Requests a single item from the table with specified id
      *
-     * @param id the id of the requested entry
-     * @return return the requested entry if found, null if not found
+     * @param id The id of the requested entry
+     * @return Return the requested entry if found, null if not found
      */
     R get(UUID id);
 
     /**
-     * requests all entries in a table
+     * Requests all entries in a table
      *
-     * @return returns a list of all entries, empty list if there are no entries
+     * @return Returns a list of all entries, empty list if there are no entries
      */
     List<R> getAll();
 }
