+++ src/main/java/com/globallogic/amcr/persistence/dao/casestudies/CaseStudyDao.java
@@ -1,120 +1,33 @@
 package com.globallogic.amcr.persistence.dao.casestudies;
 
-import com.globallogic.amcr.persistence.mapper.casestudies.CaseStudyMapper;
-import com.globallogic.amcr.persistence.dao.Dao;
+import com.globallogic.amcr.persistence.dao.CrudDao;
 import com.globallogic.amcr.persistence.model.casestudies.CaseStudy;
 import com.globallogic.amcr.persistence.model.casestudies.CaseStudyOverview;
-import com.globallogic.amcr.utils.Assert;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.UUID;
 
-@Repository
-public class CaseStudyDao implements Dao<CaseStudy, CaseStudy> {
-    private final Logger Log = LoggerFactory.getLogger(CaseStudyDao.class.getName());
-    private final CaseStudyMapper caseStudyMapper;
-
-    public CaseStudyDao(CaseStudyMapper caseStudyMapper) {
-        this.caseStudyMapper = Assert.assertNull(caseStudyMapper, "Case study mapper cannot be null");
-    }
-
-    @Override
-    public CaseStudy save(CaseStudy caseStudy, UUID caseStudyId) {
-        try {
-            caseStudy.setId(caseStudyId);
-            Log.trace("DAO saving new case study:\n{}", caseStudy);
-            caseStudyMapper.save(caseStudy);
-            return caseStudy;
-        } catch (Exception e) {
-            throw new RuntimeException("Error in CaseStudyDao - could not save case study");
-        }
-    }
-
-    @Override
-    public CaseStudy get(UUID id) {
-        Log.trace("DAO requesting case study with ID {}", id);
-        return caseStudyMapper.get(id);
-    }
-
-    @Override
-    public List<CaseStudy> getAll() {
-        Log.trace("DAO requesting all case studies");
-        return caseStudyMapper.getAll();
-    }
+public interface CaseStudyDao extends CrudDao<CaseStudy, CaseStudy> {
 
     /**
-     * requests an overview of all entries in the case_studies table
+     * Requests an overview of all entries in the case_studies table
      *
-     * @return returns a list of CaseStudyOverview objects
+     * @return Returns a list of CaseStudyOverviews
      */
-    public List<CaseStudyOverview> getAllOverviews() {
-        Log.trace("DAO requesting all case study overviews");
-        return caseStudyMapper.getAllOverviews();
-    }
+    List<CaseStudyOverview> getAllOverviews();
 
     /**
-     * requests an overview of all entries in the case_studies table where the spotlight field is true
+     * Requests an overview of all entries in the case_studies table where the spotlight field is true
      *
-     * @return returns a list of CaseStudyOverview objects
+     * @return Returns a list of CaseStudyOverview objects
      */
-    public List<CaseStudyOverview> getSpotlitOverviews() {
-        Log.trace("DAO requesting all spotlit case study overviews");
-        return caseStudyMapper.getSpotlitOverviews();
-    }
-
-
-    public List<CaseStudyOverview> getLatestOverviews(int entries) {
-        Log.trace("DAO requesting the {} most recent overviews", entries);
-        return caseStudyMapper.getLatestOverviews(entries);
-    }
+    List<CaseStudyOverview> getSpotlitOverviews();
 
     /**
-     * requests an update to a specified object in the case_studies table
+     * Requests the most recent entries in the case_studies table.
+     * The number of entries is defined by the entries parameter
      *
-     * @param newCaseStudy the new case study object with updated fields
-     * @param oldCaseStudy the old case study object requested from table
-     * @return returns the complete updated object
+     * @param entries The number of entries to be returned
+     * @return The latest {entries} entries in the database
      */
-    public CaseStudy update(UUID id, CaseStudy newCaseStudy, CaseStudy oldCaseStudy) {
-        newCaseStudy.setId(id);
-        if (oldCaseStudy.equals(newCaseStudy)) {
-            return newCaseStudy;
-        }
-        if (newCaseStudy.getTitle() == null) {
-            newCaseStudy.setTitle(oldCaseStudy.getTitle());
-        }
-        if (newCaseStudy.getOverview() == null) {
-            newCaseStudy.setOverview(oldCaseStudy.getOverview());
-        }
-        if (newCaseStudy.getCoverImageLink() == null) {
-            newCaseStudy.setCoverImageLink(oldCaseStudy.getCoverImageLink());
-        }
-        if (newCaseStudy.getBody() == null) {
-            newCaseStudy.setBody(oldCaseStudy.getBody());
-        }
-        if (newCaseStudy.getDownloadLinks() == null) {
-            newCaseStudy.setDownloadLinks(oldCaseStudy.getDownloadLinks());
-        }
-        Log.trace("DAO updating case study with ID {} and content:\n{}\n\n\n\nwith new case study:\n\n{}", id, oldCaseStudy, newCaseStudy);
-        caseStudyMapper.update(id, newCaseStudy);
-        return newCaseStudy;
-    }
-
-    /**
-     * requests deletion of an entry with specified id
-     *
-     * @param id the if of the entry to be deleted
-     */
-    public UUID delete(UUID id) {
-        try {
-            Log.trace("DAO deleting case study with ID {}", id);
-            caseStudyMapper.delete(id);
-            return id;
-        } catch (Exception e) {
-            throw new RuntimeException("Error in CaseStudyDao - could not delete case study with id " + id, e);
-        }
-    }
+    List<CaseStudyOverview> getLatestOverviews(int entries);
 }
